# openAPI.yml
openapi: 3.0.0
info:
  title: Netflix API
  description: API para acceder a la información sobre películas y programas de TV de Netflix.
  version: 1.0.0
servers:
  - url: "http://localhost:3000"
paths:
  /api/movies:
    get:
      summary: Obtiene una lista de todas las películas disponibles en Netflix.
      responses:
        "200":
          description: Lista de películas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
    post:
      summary: Añade una nueva película
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '201':
          description: Película añadida correctamente
        '400':
          description: Datos inválidos
    put:
      summary: Actualiza una película existente
      parameters:
        - in: path
          name: show_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '200':
          description: Película actualizada correctamente
        '404':
          description: Película no encontrada
        '400':
          description: Error en la solicitud
    delete:
      summary: Elimina una película
      parameters:
        - in: path
          name: show_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Película eliminada correctamente
        '404':
          description: Película no encontrada

  /api/shows:
    get:
      summary: Obtiene una lista de todos los programas de TV disponibles en Netflix.
      responses:
        "200":
          description: Lista de programas de TV.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TVShow"
  /api/movies/{show_id}:
    get:
      summary: Obtiene los detalles de una película específica.
      parameters:
        - in: path
          name: show_id
          description: Identificador único de la película.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalles de la película.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "404":
          description: Película no encontrada.
  /api/shows/{show_id}:
    get:
      summary: Obtiene los detalles de un programa de TV específico.
      parameters:
        - in: path
          name: show_id
          description: Identificador único del programa de TV.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalles del programa de TV.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TVShow"
        "404":
          description: Programa de TV no encontrado.
  /api/movies/genres/{genre}:
    get:
      summary: Obtiene una lista de películas por género.
      parameters:
        - in: path
          name: genre
          description: Género de las películas.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de películas por género.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
        "404":
          description: No se encontraron películas para el género especificado.
  /api/shows/genres/{genre}:
    get:
      summary: Obtiene una lista de programas de TV por género.
      parameters:
        - in: path
          name: genre
          description: Género de los programas de TV.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de programas de TV por género.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TVShow"
        "404":
          description: No se encontraron programas de TV para el género especificado.
  /api/movies/year/{year}:
    get:
      summary: Obtiene una lista de películas por año de lanzamiento.
      parameters:
        - in: path
          name: year
          description: Año de lanzamiento de las películas.
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Lista de películas por año de lanzamiento.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
        "404":
          description: No se encontraron películas para el año especificado.
  /api/shows/year/{year}:
    get:
      summary: Obtiene una lista de programas de TV por año de lanzamiento.
      parameters:
        - in: path
          name: year
          description: Año de lanzamiento de los programas de TV.
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Lista de programas de TV por año de lanzamiento.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TVShow"
        "404":
          description: No se encontraron programas de TV para el año especificado.
          
          
  /api/shows:
    post:
      summary: Añade un nuevo programa de TV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TVShow'
      responses:
        '201':
          description: Programa de TV añadido correctamente
        '400':
          description: Datos inválidos
    put:
      summary: Actualiza un programa de TV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TVShow'
      responses:
        '200':
          description: Programa de TV actualizado correctamente
        '404':
          description: Programa de TV no encontrado
    delete:
      summary: Elimina un programa de TV
      parameters:
        - in: query
          name: show_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Programa de TV eliminado correctamente
        '404':
          description: Programa de TV no encontrado
  /api/reviews:
    post:
      summary: Añade una nueva reseña a una película o programa de TV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '201':
          description: Reseña añadida correctamente
        '400':
          description: Datos inválidos
components:
  schemas:
    Movie:
      type: object
      properties:
        show_id:
          type: string
          description: Identificador único de la película.
        type:
          type: string
          description: Categoría de la película (Movie).
        title:
          type: string
          description: Nombre de la película.
        director:
          type: string
          description: Director(es) de la película.
        cast:
          type: string
          description: Lista de actores/actrices principales de la película.
        country:
          type: string
          description: País(es) de producción de la película.
        date_added:
          type: string
          description: Fecha en la que se añadió la película a Netflix.
        release_year:
          type: integer
          description: Año de lanzamiento de la película.
        rating:
          type: string
          description: Clasificación por edades de la película.
        duration:
          type: integer
          description: Duración de la película en minutos.
        listed_in:
          type: string
          description: Géneros de la película.
        description:
          type: string
          description: Resumen de la película.
    TVShow:
      type: object
      properties:
        show_id:
          type: string
          description: Identificador único del programa de TV.
        type:
          type: string
          description: Categoría del programa de TV (TV Show).
        title:
          type: string
          description: Nombre del programa de TV.
        director:
          type: string
          description: Director(es) del programa de TV.
        cast:
          type: string
          description: Lista de actores/actrices principales del programa de TV.
        country:
          type: string
          description: País(es) de producción del programa de TV.
        date_added:
          type: string
          description: Fecha en la que se añadió el programa de TV a Netflix.
        release_year:
          type: integer
          description: Año de lanzamiento del programa de TV.
        rating:
          type: string
          description: Clasificación por edades del programa de TV.
        duration:
          type: integer
          description: Número de temporadas del programa de TV.
        listed_in:
          type: string
          description: Géneros del programa de TV.
        description:
          type: string
          description: Resumen del programa de TV.
    User:
      type: object
      required:
        - username
        - email
      properties:
        user_id:
          type: string
          description: Identificador único del usuario
        username:
          type: string
          description: Nombre de usuario
        email:
          type: string
          description: Correo electrónico del usuario
    Review:
      type: object
      required:
        - user_id
        - show_id
        - rating
        - text
      properties:
        review_id:
          type: string
          description: Identificador único de la reseña
        user_id:
          type: string
          description: Identificador del usuario que escribe la reseña
        show_id:
          type: string
          description: Identificador de la película o programa de TV reseñado
        rating:
          type: integer
          description: Calificación dada por el usuario
        text:
          type: string
          description: Texto de la reseña